#37 Echo server using public channels is now fully supported. Private channels up next..
Fixed a bunch of things I found in the JS code, added a localDelete() function which pretty much just triggers a signal; useful for the echo server deleted events.
Fixed a small issue in the process route floor thumbnail code, if the file did not exist it'd crash, but should be ok now.

+++ app/Jobs/ProcessRouteFloorThumbnail.php
@@ -116,7 +116,9 @@ class ProcessRouteFloorThumbnail implements ShouldQueue
                 // $this->compressPng($tmpScaledFile, $target);
             } finally {
                 // Cleanup
-                unlink($tmpFile);
+                if (file_exists($tmpFile)) {
+                    unlink($tmpFile);
+                }
                 // unlink($tmpScaledFile);
                 Log::channel('scheduler')->info('Done');
             }
